requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
gitpython==3.1.40
pydantic==2.4.2
google-generativeai==0.3.1
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
pydantic-settings==2.0.0 